blkdev_max_block: take the block size as a parameter
All the callers of blkdev_max_block() actually have the block size
already, and having blkdev_max_block() look it up is actively *wrong*,
because it looks it up in the wrong place (the inode->i_blkbits field,
which may change from under it).

So just pass it in as a second parameter.

Also, fix the direct-IO block mapping function (blkdev_get_blocks()) to
be sane, and to use the buffer head size.  Again, it should not look at
i_blkbits, and if the buffer head that the direct-IO code passes extends
past the size of the block device, then we just truncate it to that size.

The old code actually did exactly that, although it was very hard to
read, because it was not at all clear what the relatonhip between b_size
and inode->i_blkbits was (normally they are related, but direct-IO is
"special", in a short-bus kind of way).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>